v.random: do not transfer cat from area if attributes are coppied too (#2537)
Current code creates two cat entries for each point on the same level -
one from area point is in and one matching new entry in the database.
As entries in db have new cat values, old cat value points to a random
or non-existing entry in the db thus causing a mess.
Fixes #2475.