Rename REQUIREMENTS.html to REQUIREMENTS.md
This has to be done in a separate commit of the actual porting, in order for git/GitHub to be able to keep track that REQUIREMENTS.html is the old file of REQUIREMENTS.md (to have a link to continue searching history). Or else, it will be a file added and a file deleted, no correlation between the two.