utils: Remove HTMLParseError (#2524)
HTMLParseError is just an alias for Exception. There is no HTMLParseError in Python 3. The error reporting fails because of access to attributes.

This completely removes HTMLParseError from HTML-man conversion code. The HTMLParseError included attributes about the error, but now the parsing is not strict anymore, so there is nothing to report.

Additionally, removes import code for Python 2.