r3.out.netcdf: Use current major version in history attribute (#2342)
r3.out.netcdf creates a netCDF file with attribute history which contains software. This removes the hard-coded software version information and uses the current major version instead for the history text. It uses fairly general macros to encapsulate stringification to avoid complex code needed to use snprintf.

The documentation contains the output, so the version should be there, but I removed it from the examples to avoid the need for updates (making the documentation little less precise).