v.perturb: Modify points for all types
This modifies points in all feature types not only type GV_POINT. While this may create unexpected results for some data, separate lines work just fine. The current code just does that always, but this should be opt-in (changes behavior) and it can be marked as experimental or just with a good description of what to expect. I test with lines, but not areas. Speed implications of the for loop not tested.