r.univar: print correct cells value for zones in shell script style (#2959)
When r.univar is called with zones and the g-flag for shell script output, the "cells" attribute gets always the number of cells in the first zone encountered, which does not correspond to the actual number of cells in the zone (or in the region in total). This corrects the value printed and adjusts the tests accordingly.