contributing: Add suggest message to version script
The helper script for udpates of versions suggests a commit message right after the update, but when additional changes need to be done, the message from the script can be lost. This adds a subcommand which can generate the message anytime. The message only makes sense when doing the version update, but the script does not check that. It also does not check (does not know) if the RC was just released. We use different commit message after that, but in that case, the commit happens right after the change from RC to dev, so there is no need to use this subcommand.