wxGUI: Rename location to project
This renames location to project in user messages in wxGUI.

This does not change any documentation or variable names. It does not change names outside of wxGUI.

This changes only location to project. It leaves mapset as is because there seems to be much higher consensus about renaming location to project than about renaming mapset in general.