CI: Reduce number of Python jobs
With increased use of pre-commit, having nice feedback from the CI checks is less important because contributors should get the feedback from pre-commit and CI is just to check, but not primarily to report to the contributor what quality controls did not pass. This applies to Black and Flake8. Pylint is not in pre-commit, but assuming the other get checked by pre-commit, the Pylint step is the one to fail most likely at the end of the job in case of issues.