CI: Reduce number of Python jobs (#3014)
With increased use of pre-commit, having nice feedback from the CI checks is less important because contributors should get the feedback from pre-commit and CI is just to check that the files are correct, but not primarily to report to the contributor what quality controls did not pass because that's what should happen locally when the contributor follows contributing guidelines.

This applies to Black and Flake8. Pylint is not in pre-commit. Assuming the other get checked by pre-commit, the Pylint step is the one to fail most likely at the end of the job in case of issues.

Black and Flake8 checks (which are purely static) are done before compilation so that generated files are not checked.

Uses matrix with only single configuration to configure versions. Reports all versions of software in one place.