contributing: Add empty line after heading in release notes (#3048)
Based on the issue reported on Gitter 'there must be an empty line after the section titles to be valid MD.' This adds an extra newline after the headings generated by the script. Some headings are generated by the GitHub API and they don't include the line. GitHub rendering works well for both versions.