d.vect.thematic: Do not show decimal places for ints
Number of minimal decimal places is fixed to 2 in the legend generated by d.vect.thematic. This adds logic which changes the number of decimal places to 0 when the column type is integer (DB_C_TYPE_INT).

This does not let user decide the number, but it covers more cases and it prepares a way for making it parametrized in the future.