tests: Display pytest failures as annotations in GitHub Actions (#3207)
Add a pytest plugin that bridges the pytest failures and special log format that is understood by GitHub Actions creating annotations in GitHub PRs. This avoids browsing the log or navigating IDE when locating the failing piece of code in question. While some Pylint-detected issues may show line 1 instead of the actual line, annotation is still helpful as a first step in learning about the detected issue.