v.to.rast: Change use=val to use=value (#3110)
The use parameter (option) had value 'val' while the parameter which is using that is called value. This makes it consistent, so use=value and value=1, not use=val and value=1. Both can be shortened, so even use=val and val=1 are still valid parameters.

This changes code which is using v.to.rast.

This does not change the other possible values (options), namely attr, cat, and dir.