pre-commit: set min-space to 1 for in-line comments in yaml files
Renovate bot commit may contain trailing comments with 1 space from
the content. This is caused by Renovate uses 'prettier' for formatting, which
has a different default compared to yamllint, and that is not likely to change.

This is completely arbitrary, thus to avoid further linting issues cased by
Renovate we can as well adapt to use 1 min-space from content.