grass.pygrass: Support shortened parameter values (#3275)
Command line interface parser allows parameter (option) values such as 'val' when full value should be 'value'. Parameter class from pygrass does the checking, but does not know about these rules. This addition covers the simple case of val-value which is not much work to implement and maintain. It does not cover more complex cases with underscores and legacy aliases.

The test covers the issue for use=value change in v.to.rast made in #3110.

Strict checking in pygrass is discussed in #3237.