lib: Use integer port argument with mysql_real_connect (#3325)
The port argument is not a string, but a plain integer.  In the past, compilers ignore such type errors and produced a warning only, but this is changing, so this change also addresses a potential build failure.

Unlike  PQsetdbLogin() which uses char *, mysql_real_connect() uses unsigned int for the port.

Found as part of Porting To Modern C in the Fedora Project.