wxGUI: Ignore stderr of the GUI (#3049)
When GUI is started by the init (grass.py) script, error messages coming from wxGUI are now ignored. The messages can be inspected by setting GRASS_DEBUG.

There is many GTK critical messages which seem very critical, but are harmless for the end user and we had little luck in fixing them, perhaps because we are using GTK only through wxPython and not directly. Complains about the messages far outnumber complains about non-functional widgets on GTK, so hiding the messages not relevant to the end user by default makes sense. Other programs typically don't face this issue as they are not started from terminal or they use daemon process for the main process, so it's unclear what is the common practice except for the result being no messages are typically visible to the end user.

* React to GRASS GIS environment WX_DEBUG besides system environment variable GRASS_DEBUG