Allow to override build date
in builds from a release tarball, there is no git commit
to take a date from, so the current date was inserted there.

See https://reproducible-builds.org/ for why this matters
and https://reproducible-builds.org/specs/source-date-epoch/
for the definition of this variable.

This patch was done while working on reproducible builds for openSUSE.