contributing: Align backporting text and example (#2978)
The first part of the paragraph talks about backporting to all branches in between main and furthest backport. However, the example pushed the idea of backport to the furthest instead. The new example is now only about backporting to the in-between branch while the paragraph addresses the backport in general and packport to the closest branch first.

The new example uses version numbers with the major version 3 because it is only an example, so it does not need to be updated.