grass.script: Provide env parameter in the g.message API
This makes the message, verbose and other functions consistent with other wrappers around run_command family calls. While not needed for multiple-mapset situations and parallelization, it is necessary when calls with messages are used without global environment being set and only custom environment available in grass.script.setup.

The PR aims at providing the interface, not updating all use cases (except the anticipated changes in grass.script.setup).

This does not have any test since the API for that is missing. This will be tested indirectly in the future.