wxGUI/history: Create panel displaying info about command (#3365)
The history log .wxgui_history is newly stored as a JSON file with command and command_info keys. When a new command is launched, it is stored as a new entry to a JSON file. The old plain text is preserved but not further used (we have two read methods but only one write method).

The command info contains the following:
runtime, timestamp, status (success, failed, aborted), if 2D/3D is set, and region settings.

The History browser is newly the object of the Splitter window. At the upper part, there is a tree catalog with history commands, and at the bottom part, there is a panel displaying the command info.


---------

Co-authored-by: lindakladivova <l.kladivova@seznam.cz>
Co-authored-by: Anna Petrasova <kratochanna@gmail.com>