wxGUI: Minimal changes to rename location to project (#2993)
This renames location to project in user messages in wxGUI.
This does not change any documentation or variable names. It does not change names outside of wxGUI.
This changes only location to project. It leaves mapset as is because there seems to be much higher consensus about renaming location to project than about renaming mapset in general.
See also https://github.com/OSGeo/grass/issues/3121

Co-authored-by: Caitlin Haedrich
Co-authored-by: Veronica Andreo <veroandreo@gmail.com>
Co-authored-by: Anna Petrasova <kratochanna@gmail.com>
Co-authored-by: Markus Neteler <neteler@osgeo.org>