grass.pygrass: Lazy load list of commands for module shortcuts
Importing _grass.pygrass.modules_ required the list to be loaded because the _shortcuts_ module is loaded by the _modules_ module. Now the list is loaded only when the _dir_ function is called for the first time.

This is using the technique from grass init file for lazy loading translations.

This allows grass.pygrass.modules to be imported without an active session, so this simplifies the GridModule tests (and its imports in general).