contributing: Add more known authors to release notes
To match the commits to their GitHub names, we need to translate some of the names and emails to GitHub names. Some of these were missing. Now the script also generates a list of authors who need that update.

There is still some names which are not recognized and we need to have another mechanism to ignore bots, so this is a draft.