contributing: Add more prefixes
Add prefixes temporal (a lot of temporal tools are often changed together). Docker and CI et al. can have subcategories or scopes. General style and perf are recognized. unittest (as in Python unittest) is recognized together with pytest. Refactoring is added as reaction to a specific change. utils is based on existing commits which are hard to categorize otherwise.