contributing: Add API to check PR titles (#3824)
This extends the API of the release notes generator to check a string (title) provided in command line against the categories provided in utils/release.yml. If there is no match, a message is printed and the process exits with non-zero return code. There is category "other changes", so non-zero may or may not mean a failure.

This does not do the checking, only provides the API to implement it.