contributing: Add more known authors to release notes (#3795)
To match the commits to their GitHub names, we need to translate some of the names and emails to GitHub names. Some of these were missing. Now the script also generates a list of authors who need that update.

Now it prefers the new Git translation list over Subversion (and we now fall back to actually showing the issues if nothing was found in Git or Subversion lists).