checks: Disable Bandit B603 check for untrusted input (#3839)
The message 'check for execution of untrusted input' is triggered by any use of subprocess regardless of the actual input, so even fixed input triggers it. Bandit issue https://github.com/PyCQA/bandit/issues/333 discusses that this is a common false positive triggered by any usage.

The pattern is common enough in our code to ignore this to avoid clutter and warning fatigue. We already ignore B404 mentioned in the issue above.