as177.c: alfnac, fix bound condition
The array r has a maximum index of 6.
Additionally, since j is an int, it can have a negative value.
This commit limits the risk of errors when accessing r.
Moreover, since r is only used within this function,
we remove the j == 1 condition and integrate it directly into the r array.