contributing: Add monthly financial supporters to release notes (#3883)
The generated release notes now contain members of supporters tier on Open Collective. We want to highlight financial contributions in the release notes. Monthly supporters seem like a fitting category in context of release. Current (draft) release notes are already updated.

The list is randomized when generated. Images are not included because that would require generating content for the missing images and making them circular (more work and circular did not work for me).

They are included under a heading at the end of highlights section, so not at the end, but not at the very beginning either.

Also: Specify number of highlight items. Add heading for new addons. Use title case. Use 'tools' not 'modules' for heading. Add requests as pip dependency to get the URL