tools: Use 4 cores as default for parallelized tools
Given the common requests for parallelization by default and 4 cores (threads, processes) being most efficient, this is changing the default for nprocs for all tools to 4.

The value can be changed in GUI settings or using g.gisenv using the NPROCS variable.

This will require review of all test of tools with nprocs to make sure they are using nprocs=1 and not the default when parallelization is not tested.