style(gui): Sort and group imports for simple changes only
This is a first part for the changes in this folder as the changes were too long and complex to do in one pass. This only kept the low risk changes, that don't require thinking much.

Uses a combination of `ruff check --output-format=concise --select I --fix gui`, `isort --profile=black gui` and `black .`