r3.showdspf: fix null pointer dereference
In the current execution, when we face an error while trying
to open a database file, as part of error log we are dereferencing
a NULL file pointer (which is returend when we try to open a file).
Ideally we should be printing the name of the database file rather
than the file pointer. This fixes the problem of null pointer
dereference. Found via cppcheck tool.

Signed-off-by: Mohan Yelugoti <ymdatta.work@gmail.com>