v.external.out: check for valid array before passing to qsort
Currently, if 'HAVE_OGR' macro is defined, as part of execution,
we sort all formats by name using qsort. But, array containing
all formats is assigned based on a conditional and if the
conditional fails, it can be NULL.

Behavior of qsort when a NULL array is provided is undefined.
To avoid getting into that situation, check if the array is
NULL before performing qsort on it.

Signed-off-by: Mohan Yelugoti <ymdatta.work@gmail.com>