display lib D_show_conversions(): %.1f to %g where values likely to be smaller than 0.0; whitespace tweaks
Display pixels or raster region rows,cols are likely be on the order of 500, while projected map units are often in the millions, so pixels/meter factor can end up as a very small number. This patch edits an unused in the production code debug helper function to not round that division into a confusing 0.0, as well as some whitespace tweaks to make it more readable.