lib/gmath: use C interface to BLAS and LAPACK libraries
Replace outdated direct use of Fortran libraries via g2c.h,  f2c.h
or private prototypes, with the C interface API and libraries
CBLAS and LAPACKE. This simplifies our own code and makes
it easier to maintain and expand.

Configure is simplified to use pkg-config to retrieve INC and
LIB info. `--with-blas` defaults to `--with-blas=cblas` where "cblas"
is the pkg-config package name; `---with-lapack` defaults to
`--with-lapack=lapacke`.