v.generalize: Initialize all of structure contents before using it
Currently, in `head` which is a `POINT_LIST` structure, we are
only initializing the next pointer to NULL, but data present
in the point substructure is not initialized which implies that
it would be filled with random data. To avoid such scenario,
initialize point coordinates to zero.

This was found using cppcheck tool.

Signed-off-by: Mohan Yelugoti <ymdatta.work@gmail.com>