v.out.ogr: Check for valid array before passing to qsort (#4278)
Currently, filling list with GDAL driver names is based on
a conditional and if the conditional fails, the list can be NULL.

We pass this list to qsort to sort the names, but behavior
of qsort with a NULL array is undefined. To avoid getting into this
situation, check if the array is NULL before performing qsort on it.

This issue was found using cppcheck tool.

Signed-off-by: Mohan Yelugoti <ymdatta.work@gmail.com>