v.vol.rst: Initialize structure contents before passing it around (#4282)
* v.vol.rst: Initialize structure contents before passing it around

`skip_point` structure is not initialized when declared. This was
found using cppcheck tool.

Only when `cv` is true, `skip_point` contents are filled and the
structure is used. cppcheck raised the issue when 'cv' is false.
Technically, we don't need skip_point in such cases and not
initializing it won't affect the execution. Nevertheless, it's a
good practice to initialize structure contents at the time of
declaration.

Signed-off-by: Mohan Yelugoti <ymdatta.work@gmail.com>

* Use correct literal for double type

Co-authored-by: Nicklas Larsson <n_larsson@yahoo.com>

---------

Signed-off-by: Mohan Yelugoti <ymdatta.work@gmail.com>
Co-authored-by: Nicklas Larsson <n_larsson@yahoo.com>