lib: Accept more newline styles in G_getl (#3853)
Use G_getl2 in G_getl to have everywhere the same behavior of supporting all newlines. Originally, G_getl2 was created to keep the behavior of G_getl. However, now, we want G_getl2 behavior everywhere.

Keeping G_getl2 for compatibility. Code can later be clean up to use only G_getl, probably at the time when G_getl2 is removed (v9).

The new test fails without the change in the library for CRLF for G_getl, but passes with the change. CR fails because it is not supported by fgets used since 88090da (#3850), so expecting failure for CR.

Return and argument types for ctypes fopen need to be set for the code to work everywhere reliably.

---------

Co-authored-by: Nicklas Larsson <n_larsson@yahoo.com>