v.in.dwg: Retire v.in.dwg infavor of v.in.redwg
v.in.dwg is linked to proprietary library 'OpenDWG toolkit'. One
needs to become a member of OpenDWG Alliance to get the needed
libraries and headers in order to compile this module.

However, OpenDWG webpage went offline around 2011 and we have
'v.in.redwg' as replacement for it. Hence retire 'v.in.dwg' module.

Modify configure.ac and platform make to remove any remnants
of opendwg in them.

Signed-off-by: Mohan Yelugoti <ymdatta.work@gmail.com>