v.in.dwg: Retire v.in.dwg infavor of v.in.redwg (#4329)
v.in.dwg is linked to proprietary library 'OpenDWG toolkit'. One needs to become a member of OpenDWG Alliance to get the needed libraries and headers in order to compile this tool.

However, OpenDWG webpage went offline around 2011 and we have 'v.in.redwg' as replacement for it (available in grass-addons). Hence retire 'v.in.dwg' tool.

Besides the tool, this completely removes OpenDWG from configure.ac and platform.

Signed-off-by: Mohan Yelugoti <ymdatta.work@gmail.com>