r.mask.status: Check mask status through a tool and function (#2390)
Instead of using low-level test of file existence with hardcoded raster path and name, this offers a new tool to retrieve status of the raster mask.

The new r.mask.status tool reports presence or absence of the 2D raster mask and provides additional details about the mask.

There is one usage of this now and that's the one for a shell prompt. The prompt no longer relies on testing the file presence with the test program, but uses a GRASS tool to find out.

The code goes out of its way to report both mask name (currently always MASK) and the underlying raster name if it is a reclassified (without rewriting the current C API). This is to mimic the existing C functions which are returning the underlying raster if MASK is a reclass. The tool and the new C API function return both preparing a way for using an arbitrary name for the mask while having the option to look at the underlying reclassified raster map.